Fixed ocasional fault on diff installer download
In certain cases when the download fails the file name is null and the
delete file faults. This change handles that case gracefully.

+++ FormDownload.cs
@@ -80,7 +80,8 @@ namespace GitForce
             // Remove temporary file if we did not complete the download
             if(!isCompleted)
             {
-                File.Delete(TargetFile);
+                if (!String.IsNullOrEmpty(TargetFile))
+                    File.Delete(TargetFile);
                 DialogResult = DialogResult.Cancel;
             }
             else
